+++ system/src/system/FalloutHooks.mjs
@@ -1,8 +1,9 @@
 import { hotbarDropHook } from "../hooks/hotbarDropHook.mjs";
+import { initiativeHooks } from "../hooks/initiativeHooks.mjs";
 import { preCreateItemHook } from "../hooks/preCreateItemHook.mjs";
 import { readyHook } from "../hooks/readyHook.mjs";
-import { renderChatMessageHook } from "../hooks/renderChatMessageHook.mjs";
 import { renderActorSheetHook } from "../hooks/renderActorSheetHook.mjs";
+import { renderChatMessageHook } from "../hooks/renderChatMessageHook.mjs";
 import { renderItemSheetHook } from "../hooks/renderItemSheetHook.mjs";
 
 export const FalloutHooks = {
@@ -11,6 +12,7 @@ export const FalloutHooks = {
 
 		const listeners = [
 			hotbarDropHook,
+			initiativeHooks,
 			preCreateItemHook,
 			readyHook,
 			renderActorSheetHook,